requirements.txt
Flask==2.3.2
flask-login==0.6.2
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0